# openapi/api_spec.yaml
openapi: 3.0.0
paths:
...
  /story/{slug}:
    get:
      operationId: StoryController_findAStory
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
    delete:
      operationId: StoryController_deleteAStory
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
info:
  title: TIP Medium Service
  description: A Medium Service Provider
  version: '1.0'
  contact: {}
...
servers: []
...